requirements.txt
requests
bs4 
facebook
random
os
tweepy
instabot
py3-pinterest
python-3.6
